refactor + message change
