+++ src/cli/commands/analyze/secrets.ts
@@ -17,35 +17,20 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-import { existsSync, readFileSync } from 'node:fs';
-import { join, relative } from 'node:path';
+import { existsSync } from 'node:fs';
+import { join } from 'node:path';
 import { spawnProcess } from '../../../lib/process';
 import type { SpawnResult } from '../../../lib/process';
 import { buildLocalBinaryName, detectPlatform } from '../../../lib/platform-detector';
 import { resolveAuth } from '../../../lib/auth-resolver';
 import logger from '../../../lib/logger';
-import { blank, error, info, print, success, text, warn } from '../../../ui';
+import { blank, error, print, success, text } from '../../../ui';
 import { CommandFailedError, InvalidOptionError } from '../_common/error.js';
 import { BIN_DIR } from '../../../lib/config-constants';
-import { SonarQubeClient } from '../../../sonarqube/client';
-import type { A3sIssue } from '../../../sonarqube/client';
-import { loadState, findExtensionsByProject } from '../../../lib/state-manager';
-import type { HookExtension } from '../../../lib/state';
 
 export interface AnalyzeSecretsOptions {
   file?: string;
   stdin?: boolean;
-}
-
-export interface AnalyzeFileOptions {
-  file: string;
-  branch?: string;
-}
-
-export interface AnalyzeA3sOptions {
-  file: string;
-  branch?: string;
-  project?: string;
 }
 
 export async function analyzeSecrets(options: AnalyzeSecretsOptions): Promise<void> {
@@ -366,226 +351,4 @@ function handleScanError(err: unknown): void {
 
   blank();
   throw new CommandFailedError(errorMessage);
-}
-
-// ---------------------------------------------------------------------------
-// Full file analysis pipeline: secrets → A3S
-// ---------------------------------------------------------------------------
-
-const SECRETS_FOUND_EXIT_CODE = 51;
-
-export async function analyzeFile(options: AnalyzeFileOptions): Promise<void> {
-  const { file, branch } = options;
-
-  if (!file) {
-    throw new InvalidOptionError('--file is required');
-  }
-
-  if (!existsSync(file)) {
-    throw new InvalidOptionError(`File not found: ${file}`);
-  }
-
-  // Step 1: secrets scan
-  const secretsResult = await runSecretsOnFile(file);
-  if (secretsResult === 'secrets-found') {
-    blank();
-    warn('Secrets detected in this file.');
-    text('  Remove the secrets before running full analysis to get additional issues.');
-    blank();
-    return;
-  }
-
-  // Step 2: A3S analysis
-  await runA3sAnalysis(file, branch);
-}
-
-// ---------------------------------------------------------------------------
-// Standalone A3S analysis
-// ---------------------------------------------------------------------------
-
-export async function analyzeA3s(options: AnalyzeA3sOptions): Promise<void> {
-  const { file, branch, project } = options;
-
-  if (!existsSync(file)) {
-    throw new InvalidOptionError(`File not found: ${file}`);
-  }
-
-  await runA3sAnalysis(file, branch, project);
-}
-
-// ---------------------------------------------------------------------------
-// Shared A3S logic
-// ---------------------------------------------------------------------------
-
-async function runSecretsOnFile(file: string): Promise<'secrets-found' | 'ok'> {
-  try {
-    const platform = detectPlatform();
-    const binaryPath = join(BIN_DIR, buildLocalBinaryName(platform));
-    if (!existsSync(binaryPath)) {
-      logger.debug('sonar-secrets binary not found, skipping secrets scan');
-      return 'ok';
-    }
-
-    const auth = await resolveAuth({});
-    const result = await runScan(binaryPath, file, auth.serverUrl, auth.token);
-    if ((result.exitCode ?? 0) === SECRETS_FOUND_EXIT_CODE) {
-      return 'secrets-found';
-    }
-    return 'ok';
-  } catch (err) {
-    logger.debug(`Secrets scan error (non-blocking): ${(err as Error).message}`);
-    return 'ok';
-  }
-}
-
-async function runA3sAnalysis(
-  file: string,
-  branch?: string,
-  explicitProject?: string,
-): Promise<void> {
-  const auth = await resolveCloudAuth(explicitProject);
-  if (!auth) return;
-
-  const projectKey = explicitProject ?? resolveA3sProjectKey();
-  if (!projectKey) return;
-
-  const fileContent = readA3sFileContent(file);
-  await callA3sApiAndDisplay(auth, projectKey, file, fileContent, branch);
-}
-
-/**
- * Resolve auth and validate that the connection is SonarQube Cloud.
- * Returns null when A3S should be silently skipped (no auth / on-premise without --project).
- * Throws CommandFailedError when --project is set but the connection is not Cloud.
- */
-async function resolveCloudAuth(
-  explicitProject: string | undefined,
-): Promise<{ serverUrl: string; token: string; orgKey: string } | null> {
-  let auth;
-  try {
-    auth = await resolveAuth({});
-  } catch {
-    logger.debug('A3S analysis skipped: failed to resolve auth');
-    return null;
-  }
-
-  if (!auth.token || !auth.orgKey || auth.connectionType === 'on-premise') {
-    if (explicitProject) {
-      throw new CommandFailedError(
-        'A3S analysis requires a SonarQube Cloud connection. Run: sonar auth login',
-      );
-    }
-    logger.debug('A3S analysis skipped: no auth, missing orgKey, or on-premise server');
-    return null;
-  }
-
-  return { serverUrl: auth.serverUrl, token: auth.token, orgKey: auth.orgKey };
-}
-
-/**
- * Look up the project key for the current directory from the agentExtensions registry.
- * Returns null when A3S should be silently skipped.
- */
-function resolveA3sProjectKey(): string | null {
-  try {
-    const state = loadState();
-    const extensions = findExtensionsByProject(state, 'claude-code', process.cwd());
-    const a3sExt = extensions.find(
-      (e): e is HookExtension => e.kind === 'hook' && e.name === 'sonar-a3s',
-    );
-
-    if (!a3sExt?.projectKey) {
-      logger.debug('A3S analysis skipped: no project key found in extensions registry');
-      if (process.stdin.isTTY) {
-        info(
-          'A3S analysis is not configured for this project. ' +
-            'Run `sonar integrate claude` to set it up, or use --project to specify the project key explicitly.',
-        );
-      }
-      return null;
-    }
-
-    return a3sExt.projectKey;
-  } catch {
-    logger.debug('A3S analysis skipped: failed to resolve extensions');
-    return null;
-  }
-}
-
-/**
- * Read file content for A3S analysis.
- * Throws CommandFailedError when the file cannot be read.
- */
-function readA3sFileContent(file: string): string {
-  try {
-    return readFileSync(file, 'utf-8');
-  } catch (err) {
-    throw new CommandFailedError(`Failed to read file: ${(err as Error).message}`);
-  }
-}
-
-/**
- * Call the A3S API and display the results.
- * Throws CommandFailedError on API failure.
- */
-async function callA3sApiAndDisplay(
-  auth: { serverUrl: string; token: string; orgKey: string },
-  projectKey: string,
-  file: string,
-  fileContent: string,
-  branch: string | undefined,
-): Promise<void> {
-  const filePath = relative(process.cwd(), file);
-  const client = new SonarQubeClient(auth.serverUrl, auth.token);
-
-  blank();
-  text('Running A3S analysis...');
-
-  try {
-    const response = await client.analyzeFile({
-      organizationKey: auth.orgKey,
-      projectKey,
-      ...(branch ? { branchName: branch } : {}),
-      filePath,
-      fileContent,
-    });
-
-    displayA3sResults(response.issues, response.errors);
-  } catch (err) {
-    logger.error(`A3S analysis failed: ${(err as Error).message}`);
-    blank();
-    error('A3S analysis failed.');
-    text(`  ${(err as Error).message}`);
-    blank();
-    throw new CommandFailedError('A3S analysis failed');
-  }
-}
-
-function displayA3sResults(
-  issues: A3sIssue[],
-  errors?: Array<{ code: string; message: string }> | null,
-): void {
-  if (errors && errors.length > 0) {
-    blank();
-    error('A3S analysis returned errors:');
-    errors.forEach((e) => {
-      text(`  [${e.code}] ${e.message}`);
-    });
-    blank();
-    return;
-  }
-
-  blank();
-  if (issues.length === 0) {
-    success('A3S analysis completed — no issues found.');
-  } else {
-    error(`A3S analysis found ${issues.length} issue${issues.length === 1 ? '' : 's'}:`);
-    blank();
-    issues.forEach((issue, idx) => {
-      const location = issue.textRange ? ` (line ${issue.textRange.startLine})` : '';
-      text(`  [${idx + 1}] ${issue.message}${location}`);
-      text(`      Rule: ${issue.rule}`);
-    });
-  }
-  blank();
 }
